+++ Casks/macasc.rb
@@ -1,6 +1,6 @@
 cask "macasc" do
   version "2.0.0"
-  sha256 "9f7d483782dfbc3b2e87f100c2f5d5e10d1517e2e5e27f11570ddcd828e991fe"
+  sha256 "2228a147523f068a3e2133d136cdce6d30b554ed4bd747e56e772e7fd919c2d3"
 
   url "https://github.com/Rian445/MacAsc/releases/download/v#{version}/Mac_ASC.dmg"
   name "Mac ASC"